requirements.txt
streamlit>=1.43.2
streamlit_extras>=0.5.5
openai>=1.66.3
langchain>=0.3.19
langchain-community>=0.3.18
langchain-text-splitters>=0.0.1
python-docx>=1.1.2
chromadb>=0.6.3
pandas>=2.0.0
openpyxl>=3.1.5
bcrypt>=4.0.1
beautifulsoup4>=4.12.2
requests>=2.31.0
pytesseract>=0.3.13
pysqlite3-binary
pymupdf
